backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
mysqlclient>=2.2
Pillow>=10.0
python-dotenv>=1.0.1
